Add Windows kernel handle enumeration library + address review feedback
- Create lib/msf/core/exploit/local/windows_kernel/handle_enum.rb
- Refactor module to use library (avoids code duplication)
- Fix documentation syntax highlighting (remove javascript)
- Remove nil default from EXPORT_CSV option
- Add format_file_size helper to library
