+++ lib/msf/core/exploit/local/windows_kernel/handle_enum.rb
@@ -1,0 +1,177 @@
+# -*- coding: binary -*-
+
+module Msf
+  module Exploit
+    module Local
+      module WindowsKernel
+        module HandleEnum
+          # Enumerate system handles using NtQuerySystemInformation
+          def enum_system_handles(session, max_handles = 50000, timeout = 30)
+            pointers = []
+            max_attempts = 5
+            attempt = 0
+            buffer_size = 1024 * 1024
+
+            begin
+              Timeout.timeout(timeout) do
+                while attempt < max_attempts
+                  vprint_status("HandleEnum: Attempt #{attempt + 1}: Trying buffer size #{buffer_size} bytes")
+
+                  begin
+                    buffer = "\x00" * buffer_size
+                  rescue ArgumentError
+                    print_error("HandleEnum: Failed to allocate buffer of size #{buffer_size}")
+                    return nil
+                  end
+
+                  result = session.railgun.ntdll.NtQuerySystemInformation(64, buffer, buffer_size, 4)
+
+                  if result.nil?
+                    print_error('HandleEnum: NtQuerySystemInformation returned nil')
+                    return nil
+                  end
+
+                  if result['return'] == 0
+                    vprint_good("HandleEnum: Success with buffer size #{buffer_size}")
+                    data = result['SystemInformation']
+                    break
+                  elsif result['return'] == 0xC0000004
+                    if result['ReturnLength'] && result['ReturnLength'] > buffer_size
+                      buffer_size = result['ReturnLength']
+                      vprint_status("HandleEnum: Buffer too small, need #{buffer_size} bytes")
+                    else
+                      buffer_size *= 2
+                      vprint_status("HandleEnum: Doubling buffer to #{buffer_size} bytes")
+                    end
+                    attempt += 1
+                  else
+                    print_error("HandleEnum: NtQuerySystemInformation failed: 0x#{result['return'].to_s(16)}")
+                    return nil
+                  end
+                end
+
+                if attempt >= max_attempts
+                  print_error("HandleEnum: Failed to get valid buffer after #{max_attempts} attempts")
+                  return nil
+                end
+
+                if data.nil? || data.length < 16
+                  print_error('HandleEnum: Invalid response data')
+                  return nil
+                end
+
+                num_handles = data[0, 8].unpack('Q').first
+                vprint_good("HandleEnum: System has #{num_handles} total handles")
+
+                if max_handles > 0 && num_handles > max_handles
+                  vprint_status("HandleEnum: Limiting to #{max_handles} handles")
+                  num_handles = max_handles
+                end
+
+                vprint_status("HandleEnum: Processing #{num_handles} handles...")
+
+                offset = 16
+                entry_size = 48
+
+                num_handles.times do |i|
+                  entry_offset = offset + (i * entry_size)
+                  break if entry_offset + entry_size > data.length
+
+                  entry = data[entry_offset, entry_size]
+
+                  begin
+                    object_ptr = entry[0, 8].unpack('Q').first
+                    pid = entry[8, 8].unpack('Q').first
+                    handle = entry[16, 8].unpack('Q').first
+                    access = entry[24, 4].unpack('V').first
+                    type_idx = entry[30, 2].unpack('v').first
+
+                    if kernel_address?(object_ptr)
+                      pointers << {
+                        address: object_ptr,
+                        pid: pid,
+                        handle: handle,
+                        access: access,
+                        type_index: type_idx
+                      }
+                    end
+                  rescue => e
+                    vprint_error("HandleEnum: Error parsing handle #{i}: #{e.message}")
+                  end
+                end
+
+                vprint_good("HandleEnum: Processed #{num_handles} handles, found #{pointers.size} kernel addresses")
+              end
+            rescue Timeout::Error
+              print_error("HandleEnum: Enumeration timed out after #{timeout} seconds")
+              return nil
+            rescue => e
+              print_error("HandleEnum: Error during enumeration: #{e.message}")
+              return nil
+            end
+
+            pointers
+          end
+
+          # Check if an address is a valid kernel address
+          def kernel_address?(addr)
+            return false if addr.nil? || addr == 0
+            high_bits = (addr >> 48) & 0xFFFF
+            high_bits == 0xFFFF
+          end
+
+          # Get object type hint from type index
+          def get_type_hint(type_idx)
+            hints = {
+              7 => 'Process',
+              8 => 'Thread',
+              16 => 'Key',
+              24 => 'File',
+              32 => 'ALPC',
+              33 => 'ALPC',
+              34 => 'ALPC',
+              35 => 'ALPC Port',
+              36 => 'ALPC Port',
+              37 => 'ALPC Port',
+              38 => 'ALPC Port',
+              39 => 'ALPC Port',
+              40 => 'ALPC Port',
+              41 => 'ALPC Port',
+              42 => 'ALPC Section',
+              43 => 'ALPC',
+              44 => 'ALPC',
+              45 => 'ALPC',
+              46 => 'ALPC',
+              47 => 'ALPC',
+              48 => 'ALPC'
+            }
+            hints[type_idx] || 'Unknown'
+          end
+
+          # Get process name from PID
+          def get_process_name(session, pid)
+            begin
+              session.sys.process.each_process do |p|
+                return p['name'] if p['pid'] == pid
+              end
+            rescue ::StandardError
+              nil
+            end
+            "PID:#{pid}"
+          end
+
+          # Format file size for display
+          def format_file_size(bytes)
+            if bytes < 1024
+              "#{bytes} bytes"
+            elsif bytes < 1024 * 1024
+              "#{(bytes / 1024.0).round(2)} KB"
+            else
+              "#{(bytes / (1024.0 * 1024.0)).round(2)} MB"
+            end
+          end
+        end
+      end
+    end
+  end
+end
